input from WP5, test on IACS_class_status integration
This is the first commit after vacations. It is a summary of all changes
which accumulated so far. It is on the requirements provided by WP5
through the xlsx template (Panagiotis) and some test data to check on
how to integrate data on a "one out of a list" restriction.